Move `Job` into a separate file
